Cleaned, removed Json, now weigted by stake.

+++ src/beez.go
@@ -7,6 +7,6 @@ type neighbourhood struct {
 
 type node struct {
 	Id       uint64
-	Earnings float64
-	stake    float64
+	Earnings int
+	stake    int
 }
