+++ src/postage.go
@@ -1,20 +1,20 @@
 package main
 
 type postageContract interface {
-	CollectWinnings(roundPrice float64, no *node)
-	GetTotalPayout() float64
+	CollectWinnings(roundPrice int, no *node)
+	GetTotalPayout() int
 }
 
 type simpleFixedPostage struct {
-	totalWithdrawn float64
+	totalWithdrawn int
 }
 
-func (sfp *simpleFixedPostage) CollectWinnings(roundPrice float64, no *node) {
+func (sfp *simpleFixedPostage) CollectWinnings(roundPrice int, no *node) {
 	// Fixed just uses price without considering storage space in network.
 	// Assumes static.
 	sfp.totalWithdrawn += roundPrice
 	no.Earnings += roundPrice
 }
-func (sfp *simpleFixedPostage) GetTotalPayout() float64 {
+func (sfp *simpleFixedPostage) GetTotalPayout() int {
 	return sfp.totalWithdrawn
 }
